% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Clean occurrences}
\usage{
clean_occurrences(archive, filters)
}
\arguments{
\item{archive}{Character. Path to the archive.}

\item{filters}{List.}
}
\value{
The status value returned by the zip command, invisibly.
}
\description{
Clean occurrence files in an archive.
}
\examples{
\dontrun{

clean_occurrences("dwca.zip", list())

}
}
