% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrences.R
\name{count_occurrences}
\alias{count_occurrences}
\title{Count occurrences}
\usage{
count_occurrences(x, ...)
}
\arguments{
\item{x}{Object to count occurrences for.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer.
}
\description{
Count the number of occurrences.
}
\examples{
 \dontrun{

count_occurrences(list(collection = "HR.3991"))

}
}
