% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_datasets.R
\name{gbif_datasets}
\alias{gbif_datasets}
\title{GBIF datasets}
\usage{
gbif_datasets(
  url = Sys.getenv("GBIF_API"),
  installation = Sys.getenv("GBIF_INSTALLATION")
)
}
\arguments{
\item{url}{Character. URL of GBIF API. Defaults to system environment
variable, "GBIF_API".}

\item{installation}{Character. ID key of GBIF installation. Defaults to
system environment variable, "GBIF_INSTALLATION".}
}
\value{
A list.
}
\description{
Get metadata for GBIF registered datasets of a given installation.
}
\examples{
gbif_datasets()

}
