% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archive_path.R
\name{get_archive_path}
\alias{get_archive_path}
\title{Get archive file path}
\usage{
get_archive_path(collection_id, dir = "archives")
}
\arguments{
\item{collection_id}{Character. Collection id.}

\item{dir}{Character. Path to the archive directory.}
}
\value{
Character. The file path of the archive.
}
\description{
Get the file path of an archive for a collection.
}
\examples{
\dontrun{

get_archive_path("HR.3991")

}
}
