% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collections.R
\name{get_collection_ids}
\alias{get_collection_ids}
\title{Get collection IDs}
\usage{
get_collection_ids()
}
\value{
A character vector.
}
\description{
Get collection IDs of FinBIF collections that are published to GBIF.
}
\examples{
get_datasets()
}
