% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Get endpoint}
\usage{
get_endpoint(collection_id, url_base = Sys.getenv("ENDPOINTS"))
}
\arguments{
\item{collection_id}{Character. ID string of FinBIF collection.}

\item{url_base}{Character. The base URL for the collection's data endpoint.
Defaults to system environment variable, "ENDPOINTS".}
}
\value{
A list.
}
\description{
Get FinBIF collection data endpoint needed for GBIF registration.
}
\examples{
\dontrun{

get_endpoint("HR.3991")

}
}
