% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(
  collection_id,
  org = Sys.getenv("GBIF_ORG"),
  installation = Sys.getenv("GBIF_INSTALLATION")
)
}
\arguments{
\item{collection_id}{Character. ID string of FinBIF collection.}

\item{org}{Character. GBIF organization key. Defaults to system environment
variable, "GBIF_ORG".}

\item{installation}{Character. ID key of GBIF installation. Defaults to
system environment variable, "GBIF_INSTALLATION".}
}
\value{
A list.
}
\description{
Get FinBIF collection metadata needed for GBIF registration.
}
\examples{
\dontrun{

get_metadata("HR.3991")

}
}
