% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrence_file.R
\name{get_occurrence_file}
\alias{get_occurrence_file}
\title{Get occurrence file name.}
\usage{
get_occurrence_file(
  filter,
  prefix = "occurrence",
  subsets = config::get("subsets")
)
}
\arguments{
\item{filter}{List.}

\item{prefix}{Character.}

\item{subsets}{List.}
}
\value{
Character. The file name holding occurrence records.
}
\description{
Get the file name of occurrences in an archive
}
\examples{
\dontrun{

get_occurrence_file(list())

}
}
