% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subsets.R
\name{get_subsets}
\alias{get_subsets}
\title{Get subsets}
\usage{
get_subsets(collection_id, subsets = config::get("subsets"))
}
\arguments{
\item{collection_id}{Character. ID string of FinBIF collection.}

\item{subsets}{List.}
}
\value{
A list.
}
\description{
Get subset filters for a collection.
}
\examples{
\dontrun{

get_subsets("HR.3991")

}
}
