% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uuid.R
\name{get_uuid}
\alias{get_uuid}
\title{Get UUID}
\usage{
get_uuid(registration)
}
\arguments{
\item{registration}{Integer.}
}
\value{
A list.
}
\description{
Get the UUID of a registered dataset.
}
\examples{
\dontrun{

registration <- get_registration(gbif_datasets(), "HR.3991")
get_uuid(registration)

}
}
