% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_registered.R
\name{is_registered}
\alias{is_registered}
\title{Check registration}
\usage{
is_registered(datasets, collection_id)
}
\arguments{
\item{datasets}{List. GBIF dataset metadata retrieved using \code{gbif_datasets}.}

\item{collection_id}{Character. ID string of FinBIF collection.}
}
\value{
A character vector.
}
\description{
Check if a FinBIF collection is registered with GBIF.
}
\examples{
\dontrun{

is_registered(gbif_datasets(), "HR.3991")

}
}
