% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_mod.R
\name{last_mod}
\alias{last_mod}
\title{Get last modified date}
\usage{
last_mod(x, ...)
}
\arguments{
\item{x}{Object to get last modified time for.}

\item{...}{Arguments passed to methods.}
}
\value{
A Date object.
}
\description{
Get the last modified data for FinBIF records
}
\examples{
\dontrun{

last_mod(list(collection = "HR.3991"))

}
}
