% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_modified.R
\name{last_modified}
\alias{last_modified}
\title{Get last modified date}
\usage{
last_modified(filter)
}
\arguments{
\item{filter}{List of named character vectors. Filters to apply to records.}
}
\value{
A Date object.
}
\description{
Get the last modified data for FinBIF records
}
\examples{
\dontrun{

last_modified(c(collection = "HR.3991"))

}
}
