% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_archived_subsets.R
\name{n_archived_subsets}
\alias{n_archived_subsets}
\title{Number of archived subsets}
\usage{
n_archived_subsets(archive)
}
\arguments{
\item{archive}{Darwin Core archive file.}
}
\value{
Integer.
}
\description{
Count the number of occurrence data subsets that have been archived.
}
\examples{
 \dontrun{

n_archived_subsets("archive.zip")

}
}
