% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_gbif_dataset_id.R
\name{send_gbif_dataset_id}
\alias{send_gbif_dataset_id}
\title{GBIF dataset identifier}
\usage{
send_gbif_dataset_id(
  id,
  uuid,
  url = Sys.getenv("GBIF_API"),
  user = Sys.getenv("GBIF_USER"),
  pass = Sys.getenv("GBIF_PASS")
)
}
\arguments{
\item{id}{Character. FinBIF collection ID for dataset.}

\item{uuid}{Character. GBIF dataset identifier. Returned by
\code{send_gbif_dataset_metadata}.}

\item{url}{Character. URL of GBIF API. Defaults to system environment
variable, "GBIF_API".}

\item{user}{Character. GBIF username. Defaults to system environment
variable, "GBIF_USER".}

\item{pass}{Character. GBIF password. Defaults to system environment
variable, "GBIF_PASS".}
}
\value{
If successful returns \code{NULL} invisibly.
}
\description{
Send FinBIF dataset identifier to GBIF.
}
\examples{
\dontrun{

m <- get_metadata("HR.3991")
uuid <- send_gbif_dataset_metadata(m)
send_gbif_dataset_id("HR.3991", uuid)

}
}
