% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_collection.R
\name{skip_collection}
\alias{skip_collection}
\title{Skip collection}
\usage{
skip_collection(skip = !config::get("enabled"))
}
\arguments{
\item{skip}{Logical.}
}
\value{
Logical.
}
\description{
Should the collection be skipped?
}
\examples{
\dontrun{

skip_collection()

}
}
