% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_gbif.R
\name{skip_gbif}
\alias{skip_gbif}
\title{Skip GBIF update}
\usage{
skip_gbif(collection_id, enabled = config::get("gbif"))
}
\arguments{
\item{collection_id}{Character. Collection id.}

\item{enabled}{Logical.}
}
\value{
Logical.
}
\description{
Should updating the collection for GBIF be skipped?
}
\examples{
\dontrun{

skip_gbif("HR.139")

}
}
