% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_archive.R
\name{stage_archive}
\alias{stage_archive}
\title{Get archive file path}
\usage{
stage_archive(archive, stage = "stage")
}
\arguments{
\item{archive}{Character. Path to the archive.}

\item{stage}{Character. Path to the staging directory.}
}
\value{
Character. The file path of the staged archive.
}
\description{
Get the file path of an archive for a collection.
}
\examples{
\dontrun{

stage_archive("archive.zip")

}
}
