% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstage_archive.R
\name{unstage_archive}
\alias{unstage_archive}
\title{Unstage archive}
\usage{
unstage_archive(staged_archive, dir = "archives")
}
\arguments{
\item{staged_archive}{Character. Path to the staged archive.}

\item{dir}{Character. Path to the archive directory.}
}
\value{
Character. The file path of the staged archive.
}
\description{
Unstage an updated archive file.
}
\examples{
\dontrun{

publish_archive("stage/archive.zip")

}
}
