% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gbif_dataset_endpoint.R
\name{update_gbif_dataset_endpoint}
\alias{update_gbif_dataset_endpoint}
\title{Update GBIF endpoint}
\usage{
update_gbif_dataset_endpoint(
  endpoint,
  uuid,
  url = Sys.getenv("GBIF_API"),
  user = Sys.getenv("GBIF_USER"),
  pass = Sys.getenv("GBIF_PASS")
)
}
\arguments{
\item{endpoint}{Character. URL of dataset endpoint generated by
\code{get_endpoint}.}

\item{uuid}{Character. GBIF dataset identifier.}

\item{url}{Character. URL of GBIF API. Defaults to system environment
variable, "GBIF_API".}

\item{user}{Character. GBIF username. Defaults to system environment
variable, "GBIF_USER".}

\item{pass}{Character. GBIF password. Defaults to system environment
variable, "GBIF_PASS".}
}
\value{
If successful returns \code{NULL} invisibly.
}
\description{
Update FinBIF dataset endpoint for GBIF.
}
\examples{
\dontrun{

m <- get_metadata("HR.3991")
ep <- get_endpoint("HR.3991")
uuid <- send_gbif_dataset_metadata(m)
update_gbif_dataset_endpoint(ep, uuid)

}
}
