% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gbif_dataset_metadata.R
\name{update_gbif_dataset_metadata}
\alias{update_gbif_dataset_metadata}
\title{Update metadata}
\usage{
update_gbif_dataset_metadata(
  metadata,
  uuid,
  created,
  url = Sys.getenv("GBIF_API"),
  user = Sys.getenv("GBIF_USER"),
  pass = Sys.getenv("GBIF_PASS")
)
}
\arguments{
\item{metadata}{List. FinBIF dataset metadata generated by \code{get_metadata}.}

\item{uuid}{Character. GBIF dataset identifier.}

\item{created}{Character. Time dataset created in ISO8601 format.}

\item{url}{Character. URL of GBIF API. Defaults to system environment
variable, "GBIF_API".}

\item{user}{Character. GBIF username. Defaults to system environment
variable, "GBIF_USER".}

\item{pass}{Character. GBIF password. Defaults to system environment
variable, "GBIF_PASS".}
}
\value{
A list.
}
\description{
Update FinBIF dataset metadata at GBIF.
}
\examples{
\dontrun{

m <- get_metadata("HR.3991")
send_gbif_dataset_metadata(m)

}
}
