% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{Write EML}
\usage{
write_eml(archive, collection_id, uuid, metadata, eml = config::get("eml"))
}
\arguments{
\item{archive}{Character. Path to a DarwinCore archive.}

\item{collection_id}{Collection ID.}

\item{eml}{List.}
}
\value{
The status value returned by the zip command, invisibly.
}
\description{
Write an EML metadata file.
}
\examples{
\dontrun{

registration <- get_registration(gbif_datasets(), "HR.3991")
uuid <- get_uuid(registration)
write_eml("dwca.zip", "HR.447", uuid, list())

}
}
