% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_occurrences.R
\name{write_occurrences}
\alias{write_occurrences}
\title{Write occurrences}
\usage{
write_occurrences(data, archive, file_name = "occurrence.txt")
}
\arguments{
\item{data}{A data.frame. Occurrence records.}

\item{archive}{Character. Path to the archive.}

\item{file_name}{Character. The name of the file to write to the archive.}
}
\value{
A finbif_occ object.
}
\description{
Write occurrence records to a Darwin Core archive.
}
\examples{
data <- get_occurrences(
  c(collection = "HR.3991"), c("occurrenceID", "basisOfRecord"), 100
)
write_occurrences(data, "dwca.zip")
}
