% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_occurrences.R
\name{archive_occurrences}
\alias{archive_occurrences}
\title{Archive occurrences}
\usage{
archive_occurrences(
  archive,
  file_name,
  media_file_name,
  filter,
  select = sub("^.*:", "", config::get("fields")),
  facts = config::get("facts"),
  combine = config::get("combine"),
  n = config::get("nmax"),
  quiet = TRUE
)
}
\arguments{
\item{archive}{Character. Path to the archive.}

\item{file_name}{Character. The name of the file to write to the archive.}

\item{media_file_name}{Character. The name of the media extension file to
write to the archive.}

\item{filter}{List of named character vectors. Filters to apply to records.}

\item{select}{Character vector. Variables to return. If not specified, a
default set of commonly used variables will be used. Use \code{"default_vars"}
as a shortcut for this set. Variables can be deselected by prepending a \code{-}
to the variable name. If only deselects are specified the default set of
variables without the deselection will be returned.}

\item{facts}{List of extra variables to be extracted from record,
event and document "facts".}

\item{combine}{List of fields to combine.}

\item{n}{Integer. How many records to download/import.}

\item{quiet}{Logical. Suppress the progress indicator for multipage
downloads.}
}
\value{
The status value returned by the zip command, invisibly.
}
\description{
Archive occurrence records in a Darwin Core archive.
}
\examples{
\dontrun{

archive_occurrences(
  "dwca.zip", "occurrence.txt", list(collection = "HR.139"),
   c("occurrenceID", "basisOfRecord")
)

}
}
