% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subsets.R
\name{get_subsets}
\alias{get_subsets}
\title{Get subsets}
\usage{
get_subsets(
  collection_id,
  filters = config::get("filters"),
  nmax = config::get("nmax"),
  facts = config::get("facts"),
  combine = config::get("combine"),
  update_freq = config::get("update_freq")
)
}
\arguments{
\item{collection_id}{Character. ID string of FinBIF collection.}

\item{filters}{List.}

\item{nmax}{Integer. Maximum allowed size of subset.}

\item{facts}{List.}

\item{combine}{List.}

\item{update_freq}{Character. Either "month" or "day".}
}
\value{
A list.
}
\description{
Get subset filters for a collection.
}
\examples{
\dontrun{

get_subsets("HR.3991")

}
}
