% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_archive.R
\name{publish_archive}
\alias{publish_archive}
\title{Get archive file path}
\usage{
publish_archive(staged_archive, filter, dir = "archives")
}
\arguments{
\item{staged_archive}{Character. Path to the staged archive.}

\item{filter}{List.}

\item{dir}{Character. Path to the archive directory}
}
\value{
Character. The file path of the staged archive.
}
\description{
Get the file path of an archive for a collection.
}
\examples{
\dontrun{

publish_archive("stage/archive.zip")

}
}
