% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_gbif_dataset_metadata.R
\name{send_gbif_dataset_metadata}
\alias{send_gbif_dataset_metadata}
\title{GBIF datasets}
\usage{
send_gbif_dataset_metadata(
  metadata,
  url = Sys.getenv("GBIF_API"),
  user = Sys.getenv("GBIF_USER"),
  pass = Sys.getenv("GBIF_PASS")
)
}
\arguments{
\item{url}{Character. URL of GBIF API. Defaults to system environment
variable, "GBIF_API".}

\item{installation}{Character. ID key of GBIF installation. Defaults to
system environment variable, "GBIF_INSTALLATION".}
}
\value{
A list.
}
\description{
Get metadata for GBIF registered datasets of a given installation.
}
\examples{
\dontrun{

m <- get_metadata("HR.3991")
send_gbif_dataset_metadata(m)

}
}
