% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_collection.R
\name{skip_collection}
\alias{skip_collection}
\title{Skip collection}
\usage{
skip_collection(
  collection_id,
  enabled = config::get("enabled"),
  collections = "collections.json"
)
}
\arguments{
\item{collection_id}{Character. Collection id.}

\item{enabled}{Logical.}

\item{collections}{Character. Path to collections.json file.}
}
\value{
Logical.
}
\description{
Should the collection be skipped?
}
\examples{
\dontrun{

skip_collection("HR.139")

}
}
