% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_meta.R
\name{write_meta}
\alias{write_meta}
\title{Write metafile}
\usage{
write_meta(archive, fields, files = "occurrence.txt", id = 1)
}
\arguments{
\item{archive}{Character. Path to the archive.}

\item{fields}{Character vector. The field names of the data files. Field
names can optionally be prepended with a namespace (one of "dwc", "dwciri",
"dc" or "dcterms") separated from the field by a ":". If no namespace is
specified, "dwc" will be assumed.}

\item{files}{Character vector. The filepaths of the data files relative to
the location of the metafile.}

\item{id}{Integer. Indicates which field can be considered the record
identifier. No ID field will be specified if \code{id} is not an integer
between 1 and the number of fields specified.}
}
\value{
The status value returned by the zip command, invisibly.
}
\description{
Write a Darwin Core archive meta data file.
}
\examples{
\dontrun{

write_meta("dwca.zip", c("occurrenceID", "basisOfRecord"))

}
}
